Updated test's files
