+++ test/include/ReconstructFromSfMData.h
@@ -8,11 +8,13 @@
 #ifndef RECONSTRUCTFROMSFMDATA_H_
 #define RECONSTRUCTFROMSFMDATA_H_
 
-#include <SfMData.h>
-#include <types_config.hpp>
-#include <ManifoldMeshReconstructor.h>
-#include <opencv2/core/core.hpp>
 #include <opencv2/calib3d/calib3d.hpp>
+#include <opencv2/core/core.hpp>
+
+#include <manifoldReconstructor/ManifoldMeshReconstructor.h>
+#include <manifoldReconstructor/SfMData.h>
+#include <manifoldReconstructor/types_config.hpp>
+
 
 class ReconstructFromSfMData {
 public:
