+++ test/src/EdgePointConfigurator.cpp
