+++ test/src/KittiRectifiedCamParser.cpp
